Update addon.xml
 Export lib so other addons can access youtube-dl and yt-dlp.

Normally, it would be cleaner to separate this addon's code (service.py) from shared library code (which the lib directory exclusively contains).  However, since this addon's code is contained in a single short file (which is not exported) AND the exported directory (lib) completely contains only the shared external code, then it seems it may be a bit much to split this into two addons.