Retain Python 2 compatibility (#131)
* Retain Python 2 compatibility

* Indent got added in Python 3.9

Pretty-printing isn't really needed, so don't do it if it can't be had.

* Account for module packaging 

Python 3 groups server modules into a dedicated 'http' package. Account for the old Python 2 structure dynamically.

* Remove Python 3 gate/conditional

The DASH manifest setting currently does nothing if you're on Python 2, so there's no need for this conditional.

* Remove options invalid for Youtube-dl

Youtube-dl can't handle it and yt-dlp appears to be be working without it just fine.

* Resurrect Kodi 18/19 support

* Python 2 is mostly needed by Kodi 18/19 users
* The conditional is trivial enough to bring it back, so let's do :)
* Reverts 5e484af and 4bad1e9

* Adjust dependencies' minimum versions to match latest Python 2 releases

* Adjust CI for Leia

* Fixed indentation

* No need to create existing folder